Completed extensive debugging
Run the parser on the entire sgf_store/baduk-pro-collection folder.
Revealed several problem with the parsing methods used.
sgf_parser was extensiving changed with new replacements being made and a careful ordering of the operations.
Error handling in several functions was improved to aid in finding malformed sgf files in future.


Former-commit-id: 92d72db3df3dee8da8ae55a312d7d9510d9c40e2