Implement union as __setitem__
This completes the encasing of all the logic of the unionfind into the
index syntax. It is slightly against the spirit of the union function,
as a target is selected and everything is unioned to that, but this
allows the use case to be strategic in target selection ie for unioning
to the largest group or other considerations.