Call unionfind.getitem before starting array.setitem
It seems that when the array.__setitem__ code is run, the self is fixed
in the function name space before the unionfind.__getitem__ is run, so
the additional pointers getitem can add do not make it to the
array.__setitem__ copy. The fix is to complete the getitem step
explicitly before the setitem step.