Fix the exception caused by making a move in either panel.
The kivy ids only apply to the tree explicitly written in the kv language file.
When the info button was moved into the Play Panel kv tree from the GoFamiliar kv tree,
the id no longer appeared in the list of ids for the GoFamiliar.
This caused a KeyError on the root.ids dictionary.