Refactor the on_gamestate update method
The for loop was quite long, with a large section of nearly repeated code.
Trying to read it again was difficult, so a refactor was needed to simplify it and get rid of the repetition.
It was only a double repeat, instead of the usual triple repeat to rtigger a refactor,
but the repetition was diminishing clarity.