The analysis process passes intersection search scores to the analysis board
This required completely rethinking how the app was going to work,
and adding in a great deal I initially didn't think would be nessecary.
First, multi processing is required to run the MCTS algorithm in another process to the kivy process,
so the root node is not "available" to the kivy process for reading.
Second, info passing between processes required a new function like move_search,
but that incorporated a mp.queue to pass the search scores across.
Third, both the tree policy function and the score function had to be made callable from more places,
so each was "promoted" a level.

TODO: 	use heat colours