Added heat colouring based on search scores
For each intersection, the current score of it's search is sent across the queue back to the UI,
and it is translated to a color using a pyplot cmap.
The alpha is controlled by the score itself, so a more interesting point is "hotter" and less transparent.
The map is currently set to the cool pallette as it is high contrast to the board image.

Some weird things are happening with the search algorithm, and that may take some analysis.
When the AMAf rates are used for scoring, the average score drops to zero.
When the confidence interval score is used, any searched node keeps it's high score.
There are obvious misfirings of the algorithm, and it will take some tuning to get this part right.
Perhaps sigopt or a CNN will help. Maybe bringing in the sgf library will help this out also.