Pass new moves to the analysis process
A new move in the gamne board side is passed over to the analysis side now.
The analysis is restarted on the new game state.

The main learning was the use of canvas.after on the button widget.
Before the clear command were erasing the stone image also.
Now, the canvas.after.clear command reveals the stone.

The other learning was that the game state info if hard to pass among the widgets.
I needed a long widget-tree walk to pass the state around.