Start Go Familiar starting state to be an sgf game
A new parsing function was created from the logic of several parsing functions in the sgf.read module.
I found the code to be rather awful, and some major refactoring will be needed.

Using the new import path for sgf, the new parsing function was imported,
and a hard coded example sgf is parsed and used to set the opening position.

Some problems remain, like the board is not visible until after a move.