Attempt to tune the algorithm to find good moves while doing some clean up
The colour reported by the Go Familiar is uniform after a few seconds.
I thought that tuning the parameters might make for a more "interesting" display
ie some moves show up as favoured over others.
But no, that is not the case. All moves are uniformly recommended in the display.
Some major back tracking of the algorithm will be required to chase this problem out.