Create MCTS package
Now that I am adding a neural net as a policy net, it makes sense to seperate
functions within packages under src.
This is a move to place all the old MCTS code into a single package.
Test still need to be run on this new structure.
I'll have import errors for a few days when running tests.