Some clean up to get tests to pass
The structure of the project has changed since I was last
actively working on it.
So many of the pointers for files no longer make sensae to me,
so the test won't run yet.
I am also not sure how I want to setup the data as yet, so I don't
have anything to implment yet.