Remove the .sgf extension explicitly
During the creation of the sgf pro game library, I was using the file
names without the extension.
At least I aimed to do that.
I was assuming there are no periods in the file names other than the
extension sperator.
However, the names occasionally have other periods.
This led to some weird errors, that are solved by explicitly removing
just the .sgf extension with a replace.