Rewrite the library access test
The test was formed under old assumptions about the format of the library
file which no longer hold true.
I think I hadn't recalculated the file for a long time, so the tests kept
passing even though the functions to create it had changed.

The biggest change is the addition of some spot checks of the data entered
into the library h5 file.
These are 6 specific games choosen at random where I calculated a few features
of them to test against the data stored in the library file.