Add handicap parsing to sgf.read
I had been ignoring handicaps in the sgf files.
The moves in the sgf games depended on there being handicap stones on the
board, but they were mising.
That caused many move errors to be reported.
Part of the fix was done in the thick_goban package where I enabled handicap
stones to be specified at the start.
The changes here use that fix to add the handi stones declared in the sgf file.