Debug the usage of hanicap moves
In the sgf_position case, I hadn't invoked the handicap stone argument.
In the case of create_pro_hdf5, I needed to handle the case of a handicap
move tt, which if read naively is a move off the board.

I decided to handle it as the openai gym does, where the pass move is the
next move off of the board, so 19^2 due to zero being a move on the board.