Intmove should return the move colour as well
It only passed the integer version of the move before.
This made interpreting various parts of an sgf file difficult.
By pulling through the player colour, it allows a better reading of an sgf
file as things like passes ensure that the colour of player moves do not
always alternate, so the move node colour recorded in the sgf has to be used
to play the right colour stone in the games representation.