Factor out the sgf store tests
I started to add two new test functions, and I realized I was testing the data
with these tests more than the functioanlity and correctness of the code.
As such, they needed to be in their own module to be easily understood as that
by me in the future.

The two new tests are test_pro_games_dataleaves and
test_move_count_to_position_lengths.

The changes to the notebook are the exploration I did to find out what was
causing the error I was seeing that required those two new tests.