Move sgf Library doctests to tests module
The way I handled the doctests was very clunky, with a test fixture in the actual module.
This change moves all those tests out of doctests to a test module.
Some of the docstest moved unchanged, but most needed an overhaul or were just deleted.
The test file location's contants were moved also.