Ensure bad sgf files aren't add to store
Some of the sgf files have errors in them that cause the game playouts to make illegal
moves, like playing on another stone.
There are 36 such games in the original sgf pro game store.
These changes ensure that if a game play out does not finish without errors, that game
is not added to the H5 storage file.
The test of how many files were added had to be updated also, so the bad sgf files are
excluded.